backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose
passlib[bcrypt]
python-multipart